ADD grpc client package for catalog sources
Delete all grpc code and imported from operator-registry
Moved k8s client to operator package for now.
Deleted test get node function that is no longer needed.
